+++ WireGuardDemo.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,6 +1,15 @@
 {
   "pins" : [
     {
+      "identity" : "connectorkit",
+      "kind" : "remoteSourceControl",
+      "location" : "ssh://git@git.monitorapp.com:122/aioncloud/connectorkit.git",
+      "state" : {
+        "branch" : "main",
+        "revision" : "27b1a1cc38f12e3d837d0162dba1f3f67d9b1a75"
+      }
+    },
+    {
       "identity" : "swiftybeaver",
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/SwiftyBeaver/SwiftyBeaver.git",
